Backend/requirements.txt
python-dotenv
fastapi
uvicorn
motor
pydantic